refactor: consolidate mdns module API
