+++ packages/opencode/src/mdns/client.ts
@@ -1,5 +1,8 @@
 import { Log } from "@/util/log"
-import { Bonjour, type Service, type Browser } from "bonjour-service"
+import { Bonjour, type Service } from "bonjour-service"
+
+const log = Log.create({ service: "mdns.client" })
+type Timer = ReturnType<typeof setTimeout>
 
 export interface DiscoveredServer {
   name: string
@@ -9,50 +12,45 @@ export interface DiscoveredServer {
   txt: Record<string, string>
 }
 
-export namespace MDNS {
-  const log = Log.create({ service: "mdns-client" })
-  type Timer = ReturnType<typeof setTimeout>
+export function find(timeout = 5000, quietPeriod = 1000): Promise<DiscoveredServer[]> {
+  return new Promise((resolve) => {
+    const bonjour = new Bonjour()
+    const servers: DiscoveredServer[] = []
+    let quietTimer: Timer | undefined
 
-  export function find(timeout = 5000, quietPeriod = 1000): Promise<DiscoveredServer[]> {
-    return new Promise((resolve) => {
-      const bonjour = new Bonjour()
-      const servers: DiscoveredServer[] = []
-      let quietTimer: Timer | undefined
+    const finish = () => {
+      if (quietTimer) clearTimeout(quietTimer)
+      quietTimer = undefined
+      browser.stop()
+      bonjour.destroy()
+      log.info("discovery complete", { count: servers.length })
+      resolve(servers)
+    }
 
-      const finish = () => {
-        if (quietTimer) clearTimeout(quietTimer)
-        quietTimer = undefined
-        browser.stop()
-        bonjour.destroy()
-        log.info("discovery complete", { count: servers.length })
-        resolve(servers)
-      }
+    const onService = (service: Service) => {
+      if (!service.name.startsWith("opencode-")) return
 
-      const onService = (service: Service) => {
-        if (!service.name.startsWith("opencode-")) return
+      const host = service.host
+      const port = service.port
+      const txt = service.txt ?? {}
+      const fullUrl = `http://${host}:${port}`
 
-        const host = service.host
-        const port = service.port
-        const txt = service.txt ?? {}
-        const fullUrl = `http://${host}:${port}`
+      servers.push({
+        name: service.name,
+        host,
+        port,
+        fullUrl,
+        txt,
+      })
 
-        servers.push({
-          name: service.name,
-          host,
-          port,
-          fullUrl,
-          txt,
-        })
+      log.debug("discovered server", { name: service.name, host, port })
 
-        log.debug("discovered server", { name: service.name, host, port })
+      if (quietTimer) clearTimeout(quietTimer)
+      quietTimer = setTimeout(finish, quietPeriod)
+    }
 
-        if (quietTimer) clearTimeout(quietTimer)
-        quietTimer = setTimeout(finish, quietPeriod)
-      }
+    const browser = bonjour.find({ type: "http", protocol: "tcp" }, onService)
 
-      const browser = bonjour.find({ type: "http", protocol: "tcp" }, onService)
-
-      setTimeout(finish, timeout)
-    })
-  }
+    setTimeout(finish, timeout)
+  })
 }
